+++ CHANGES.md
@@ -3,6 +3,7 @@
 #### 2.11.0 (To be released)
 
 * Upgrade to Spring Boot 2.3.0 / Spring Framework 5.2.6
+* Add [springboot-authentication-jar](examples/springboot-authentication-jar) example to demonstrate how to customise `hawtconfig.json` with Spring Boot jar app.
 
 #### 2.10.0